Imported upstream version '1.3.4' of 'upstream'

+++ osrf_testing_tools_cpp/test/test_runner/assert_env_vars.cpp
@@ -91,7 +91,7 @@ int main(int argc, char * argv[])
 {
   ::testing::InitGoogleTest(&argc, argv);
 
-  g_args.reserve(static_cast<size_t>(argc));
+  g_args.reserve(argc);
   for (int i = 0; i < argc; ++i) {
     if (i == 0) {
       continue;
